Refactor Web Speech polyfill out into environment
