+++ packages/integration-test/jest.config.json
@@ -1,4 +1,5 @@
 {
+  "testEnvironment": "./JestEnvironmentHappyDOMWithMockedWebSpeech.js",
   "testMatch": ["**/__tests__/**/*.?([cm])[jt]s?(x)", "**/?(*.)+(spec|test).?([cm])[jt]s?(x)"],
   "transform": {
     "\\.cjsx?$": [
