+++ packages/integration-test/package.json
@@ -49,7 +49,6 @@
     "@happy-dom/jest-environment": "^16.7.2",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/react": "^16.2.0",
-    "@testing-library/user-event": "^14.6.1",
     "@tsconfig/strictest": "^2.0.5",
     "@types/jest": "^29.5.13",
     "@types/react": "^18.3.11",
