+++ packages/integration-test/tsconfig.json
@@ -1,6 +1,7 @@
 // This configuration file is for VSCode only.
 {
   "compilerOptions": {
+    "noEmit": true,
     "jsx": "react"
   },
   "extends": "./tsconfig.custom.json"
